requirements.txt
# Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0

# Azure SDK
azure-identity>=1.14.0
azure-storage-file-datalake>=12.13.0

# Authentication
msal>=1.24.0
PyJWT>=2.8.0

# HTTP Requests
requests>=2.31.0

# Data Validation
great-expectations>=0.18.0

# Visualization
graphviz>=0.20.1

# Configuration Management
python-dotenv>=1.0.0

# File Format Support
openpyxl>=3.1.0
xlrd>=2.0.1
